+++ include/container/shm_vector.h
@@ -130,7 +130,7 @@ private:
 	size_t GetSuitableCapacity(size_t size) {
 		if (size < 1)
 			size = 1;
-		return util::Utility::NextPowOf2(size);
+		return util::Utility::NextPowOf2(uint32_t(size));
 	}
 
 	void shrink_to_fit() {
